refactor: refactor main TS code
