+++ src/useScreenGuard.ts
@@ -1,8 +1,11 @@
 import { useEffect, useState } from 'react';
-import { NativeEventEmitter } from 'react-native';
-import NativeScreenGuard from './NativeScreenGuard';
+import { NativeEventEmitter, NativeModules, TurboModuleRegistry } from 'react-native';
+import { type Spec } from './NativeScreenGuard';
 import * as ScreenGuardConstants from './constant';
 import { ScreenGuardHookData } from './data';
+
+const NativeScreenGuard = TurboModuleRegistry.get<Spec>('ScreenGuard')
+    || NativeModules.ScreenGuard;
 
 export function useScreenGuard() {
     const [value, setValue] = useState<ScreenGuardHookData | null>(null);
